chore: move get by id method to providers.profile
