+++ src/lib/modules/accounts/routes/resolvers/index.ts
@@ -3,6 +3,5 @@ export { getVerificationEmailStatusResolver } from './get-verification-email-sta
 export { sendEmailVerificationResolver } from './send-email-verification';
 export { getUserDetailsByIdResolver } from './get-user-details-by-auth0-id';
 export { getPracticeByUserIdResolver } from './get-practice-by-user-id';
-export { getProviderProfileByUserIdResolver } from './get-provider-profile-by-user-id';
 export { handlePracticeOnboardingResolver } from './handle-practice-onboarding';
 export { registerMemberResolver } from './register-member';
