--- src/lib/modules/accounts/service/service/get-provider-profile-by-user-id/index.ts
@@ -1,1 +1,0 @@
-export * as GetProviderProfileByUserId from './getProviderProfileByUserId';
