+++ src/lib/modules/providers/features/profiles/get-provider-profile-by-user-id/trpc.ts
@@ -1,1 +1,1 @@
-export const ROUTE = 'users.get-provider-profile-by-user-id' as const;
+export const ROUTE = 'profiles.get-profile-by-user-id' as const;
