+++ src/lib/modules/providers/routes/resolvers/index.ts
@@ -1,1 +1,2 @@
 export { resolve as listPracticeProfilesByUserIdResolver } from './list-practice-profiles-by-user-id';
+export { getProviderProfileByUserIdResolver } from './get-provider-profile-by-user-id';
