+++ src/lib/modules/providers/routes/router.ts
@@ -1,11 +1,22 @@
 import * as trpc from '@trpc/server';
 
-import { listPracticeProfilesByUserIdResolver } from './resolvers';
+import {
+    listPracticeProfilesByUserIdResolver,
+    getProviderProfileByUserIdResolver,
+} from './resolvers';
 import { Context } from '@/lib/server/context';
-import { ListPracticeProfilesByUserId } from '@/lib/modules/providers/features/profiles';
+import {
+    GetProviderProfileByUserId,
+    ListPracticeProfilesByUserId,
+} from '@/lib/modules/providers/features/profiles';
 
 export const router = trpc
     .router<Context>()
+    .query(GetProviderProfileByUserId.TRPC_ROUTE, {
+        input: GetProviderProfileByUserId.inputSchema,
+        output: GetProviderProfileByUserId.outputSchema,
+        resolve: getProviderProfileByUserIdResolver,
+    })
     .query(ListPracticeProfilesByUserId.TRPC_ROUTE, {
         input: ListPracticeProfilesByUserId.inputSchema,
         output: ListPracticeProfilesByUserId.outputSchema,
