+++ src/lib/modules/providers/service/profiles/get-profile-by-user-id/index.ts
@@ -1,0 +1,1 @@
+export * as GetProfileByUserId from './getProfileByUserId';
